+++ packages/app/src/App.tsx
@@ -45,6 +45,7 @@ import {
   GSFeatureEnabled,
   GSProviderSettings,
   GSIntegrationsLayout,
+  GSDeploymentDetailsPickerFieldExtension,
 } from '@internal/plugin-gs';
 
 const app = createApp({
@@ -133,6 +134,7 @@ const routes = (
     >
       <ScaffolderFieldExtensions>
         <GSClusterPickerFieldExtension />
+        <GSDeploymentDetailsPickerFieldExtension />
         <GSTemplateStringInputFieldExtension />
       </ScaffolderFieldExtensions>
       <ScaffolderLayouts>