+++ plugins/gs-common/src/api/types.ts
@@ -1,5 +1,6 @@
 import * as applicationv1alpha1 from '../model/applicationv1alpha1';
 import * as authorizationv1 from '../model/authorizationv1';
+import * as awsv1beta1 from '../model/awsv1beta1';
 import * as capiv1beta1 from '../model/capiv1beta1';
 import * as helmv2beta1 from '../model/helmv2beta1';
 import * as metav1 from '../model/metav1';
@@ -18,6 +19,8 @@ export type HelmRelease = helmv2beta1.IHelmRelease;
 export type List<T> = metav1.IList<T>;
 
 export type Organization = securityv1alpha1.IOrganization;
+
+export type ProviderConfig = awsv1beta1.IProviderConfig;
 
 export type Resource<T> = T & {
   installationName: string;