+++ plugins/gs-common/src/model/awsv1beta1/key.ts
@@ -1,0 +1,7 @@
+export const providerConfigApiVersion = 'aws.upbound.io/v1beta1';
+
+export const providerConfigGVK = {
+  apiVersion: 'v1beta1',
+  group: 'aws.upbound.io',
+  plural: 'providerconfigs',
+};