+++ plugins/gs/src/components/hooks/useProviderConfigs.ts
@@ -1,0 +1,7 @@
+import type { ProviderConfig } from '@internal/plugin-gs-common';
+import { providerConfigGVK } from '@internal/plugin-gs-common';
+import { useListResources } from './useListResources';
+
+export function useProviderConfigs(installations?: string[]) {
+  return useListResources<ProviderConfig>(providerConfigGVK, installations);
+}