+++ plugins/gs/src/components/scaffolder/DeploymentDetailsPicker/index.ts
@@ -1,0 +1,2 @@
+export { DeploymentDetailsPicker } from './DeploymentDetailsPicker';
+export { DeploymentDetailsPickerSchema } from './schema';