+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add experimental "Installations" page, visible with feature flag `show-installations-page`.
+- GS plugin: Add scaffolder custom field extension for picking deployment details.
 
 ## [0.31.0] - 2024-07-30